# cuda4dnn: improve host-device transfer performance