--- a/modules/dnn/src/op_cuda.hpp
+++ b/modules/dnn/src/op_cuda.hpp
@@ -13,6 +13,7 @@
 #include "cuda4dnn/csl/memory.hpp"
 #include "cuda4dnn/csl/fp16.hpp"
 #include "cuda4dnn/csl/workspace.hpp"
+#include "cuda4dnn/kernels/fp_conversion.hpp"
 #endif
 
 #include <opencv2/dnn/shape_utils.hpp>
@@ -149,7 +150,6 @@ namespace cv { namespace dnn {
             if (temp.data != destMat.data)
                 temp.copyTo(destMat);
         }
-
     }} /* namespace cuda4dnn::csl */
 
     /** base class for CUDA operation nodes (for all supported targets) */
@@ -219,6 +219,45 @@ namespace cv { namespace dnn {
         virtual void setStream(cuda4dnn::csl::Stream stream) noexcept = 0;
     };
 
+    namespace cuda4dnn { namespace detail {
+
+        template <class U>
+        void convert_D2H(const cv::Mat& mat, cuda4dnn::csl::View<U> view, cuda4dnn::csl::ManagedPtr<float>& device_temp, const cuda4dnn::csl::Stream& stream);
+
+        template <> inline
+        void convert_D2H<half>(const cv::Mat& mat, cuda4dnn::csl::View<half> view, cuda4dnn::csl::ManagedPtr<float>& device_temp, const cuda4dnn::csl::Stream& stream) {
+            if (device_temp.size() < view.size())
+                device_temp.reset(view.size());
+            auto temp_span = cuda4dnn::csl::Span<float>(device_temp.get(), view.size());
+
+            cuda4dnn::kernels::fp16_to_fp32(stream, temp_span, view);
+            cuda4dnn::csl::memcpy<float>(reinterpret_cast<float*>(mat.data), temp_span.data(), view.size(), stream);
+        }
+
+        template <> inline
+        void convert_D2H<float>(const cv::Mat& mat, cuda4dnn::csl::View<float> view, cuda4dnn::csl::ManagedPtr<float>& device_temp, const cuda4dnn::csl::Stream& stream) {
+            cuda4dnn::csl::memcpy<float>(reinterpret_cast<float*>(mat.data), view.data(), view.size(), stream);
+        }
+
+        template <class U>
+        void convert_H2D(cuda4dnn::csl::Span<U> span, const cv::Mat& mat, cuda4dnn::csl::ManagedPtr<float>& device_temp, const cuda4dnn::csl::Stream& stream);
+
+        template <> inline
+        void convert_H2D<half>(cuda4dnn::csl::Span<half> span, const cv::Mat& mat, cuda4dnn::csl::ManagedPtr<float>& device_temp, const cuda4dnn::csl::Stream& stream) {
+            if (device_temp.size() < span.size())
+                device_temp.reset(span.size());
+            auto temp_span = cuda4dnn::csl::Span<float>(device_temp.get(), span.size());
+
+            cuda4dnn::csl::memcpy<float>(temp_span.data(), reinterpret_cast<float*>(mat.data), span.size(), stream);
+            cuda4dnn::kernels::fp32_to_fp16(stream, span, temp_span);
+        }
+
+        template <> inline
+        void convert_H2D<float>(cuda4dnn::csl::Span<float> span, const cv::Mat& mat, cuda4dnn::csl::ManagedPtr<float>& device_temp, const cuda4dnn::csl::Stream& stream) {
+            cuda4dnn::csl::memcpy<float>(span.data(), reinterpret_cast<float*>(mat.data), span.size(), stream);
+        }
+    }} /* namespace cuda4dnn::detail */
+
     template <class T, int TargetID>
     class GenericCUDABackendWrapper final : public CUDABackendWrapper {
     public:
@@ -283,8 +322,12 @@ namespace cv { namespace dnn {
                  * We use a view to ensure that only the required region of memory is copied.
                  */
                 auto view = tensor_view_type(shared_block->device.get(), std::begin(shape), std::end(shape));
-                cuda4dnn::csl::copyTensorToMat<T>(view, shared_block->host, shared_block->stream);
 
+                auto& mat = shared_block->host;
+                CV_Assert(mat.isContinuous());
+                CV_Assert(mat.type() == CV_32F);
+
+                cuda4dnn::detail::convert_D2H<T>(mat, view, shared_block->device_temp, shared_block->stream);
                 shared_block->stream.synchronize();
             }
         }
@@ -300,7 +343,12 @@ namespace cv { namespace dnn {
                 shared_block->device_dirty = false;
 
                 auto span = tensor_span_type(shared_block->device.get(), std::begin(shape), std::end(shape));
-                cuda4dnn::csl::copyMatToTensor<T>(shared_block->host, span, shared_block->stream);
+
+                auto& mat = shared_block->host;
+                CV_Assert(mat.isContinuous());
+                CV_Assert(mat.type() == CV_32F);
+
+                cuda4dnn::detail::convert_H2D<T>(span, mat, shared_block->device_temp, shared_block->stream);
             }
         }
 
@@ -368,6 +416,7 @@ namespace cv { namespace dnn {
             cuda4dnn::csl::MemoryLockGuard memGuard; /* keeps host memory page-locked if possible */
 
             cuda4dnn::csl::ManagedPtr<T> device;
+            cuda4dnn::csl::ManagedPtr<float> device_temp; /* use for conversions */
             cuda4dnn::csl::Stream stream;
         };