### This pullrequest changes
- performs fp conversions on GPU
- eliminates the costly intermediate host memory allocation required for every transfer involving fp16 data

<cut/>

```
Test Size = number of 32-bit floats

Test Size: 100
	CPU Conversion H2D Time: 0.0088ms
	GPU Conversion H2D Time: 0.0148ms
	CPU Conversion D2H Time: 0.0077ms
	GPU Conversion D2H Time: 0.0098ms
Test Size: 1000
	CPU Conversion H2D Time: 0.0068ms
	GPU Conversion H2D Time: 0.0147ms
	CPU Conversion D2H Time: 0.0078ms
	GPU Conversion D2H Time: 0.0098ms
Test Size: 10000
	CPU Conversion H2D Time: 0.0125ms
	GPU Conversion H2D Time: 0.0187ms
	CPU Conversion D2H Time: 0.0140ms
	GPU Conversion D2H Time: 0.0123ms
Test Size: 100000
	CPU Conversion H2D Time: 0.0622ms
	GPU Conversion H2D Time: 0.0496ms
	CPU Conversion D2H Time: 0.0780ms
	GPU Conversion D2H Time: 0.0423ms
Test Size: 1000000
	CPU Conversion H2D Time: 0.4920ms
	GPU Conversion H2D Time: 0.3530ms
	CPU Conversion D2H Time: 0.6348ms
	GPU Conversion D2H Time: 0.3262ms
Test Size: 10000000
	CPU Conversion H2D Time: 8.5243ms
	GPU Conversion H2D Time: 3.5249ms
	CPU Conversion D2H Time: 7.6002ms
	GPU Conversion D2H Time: 3.1775ms
```

It's slower to perform the conversion on GPU for small data sizes instead of converting on CPU and then transfering. But the difference is in the order of few microseconds to few tens of microseconds. However, for large data sizes, the GPU beats CPU by a large margin.

```
force_builders=Custom,docs
buildworker:Custom=linux-4
docker_image:Custom=ubuntu-cuda:18.04

build_image:Custom Mac=openvino-2019r3.0
build_image:Custom Win=openvino-2019r3.0
test_opencl:Custom Win=OFF
test_modules:Custom Mac=dnn,java,python3
```