@YashasSamaga 2020-01-05
any news about this PR? It seems to halve inference time on some models.